mv as per PSR-4, first steps resolving incompatibilities
