+++ src/Command/All.php
