+++ tests/Command/AllTest.php
