+++ tests/Command/BuildSandboxTest.php
