+++ tests/Command/FunctionalTest.php
