+++ tests/Command/TestWatcherTest.php
