+++ tests/FunctionalTest1Test.php
@@ -13,7 +13,7 @@ class FunctionalTest1Test extends BaseTest
     {
         $this->assertTrue(true);
 
-        $databases = parse_ini_file(realpath(__DIR__.'/../../../bin/config/databases_test.ini'));
+        $databases = parse_ini_file(realpath(__DIR__.'/../bin/config/databases_test.ini'));
 
         foreach ($databases as $database) {
             $pdo = new PDO(sprintf('mysql:host=localhost;dbname=%s', $database), 'root', null);
